+++ lib/main.dart
@@ -5,6 +5,7 @@ import 'pages/StoreHoursPage.dart';
 import 'pages/settings_page.dart';
 
 import 'widgets/fun_appbar.dart';
+import 'widgets/red_appbar.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,53 +13,6 @@ void main() {
 
 
 // 通知ボタンがないAppBar, ホームに戻る
-class RedAppBar extends StatelessWidget implements PreferredSizeWidget {
-  RedAppBar(this.title);
-  final String title;
-
-  @override
-  Size get preferredSize => const Size.fromHeight(56.0);
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.grey[300],
-        borderRadius: BorderRadius.circular(12),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.white,
-            offset: Offset(-5, -5),
-            blurRadius: 15,
-          ),
-          BoxShadow(
-            color: Colors.black.withOpacity(0.3),
-            offset: Offset(5, 5),
-            blurRadius: 15,
-          )
-        ],
-      ),
-      child: AppBar(
-        title: Text(
-          title,
-          style: TextStyle(color: Colors.white),
-        ),
-        backgroundColor: Color.fromARGB(255, 210, 15, 1),
-        //あってるかわからない↓
-        actions: [],
-        leading: IconButton(
-          icon: Icon(
-            Icons.arrow_back,
-            color: Colors.white,
-          ),
-          onPressed: () {
-            Navigator.pop(
-                context, MaterialPageRoute(builder: (context) => MyHomePage()));
-          },
-        ),
-      ),
-    );
-  }
-}
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
