+++ lib/widgets/red_appbar.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+class RedAppBar extends StatelessWidget implements PreferredSizeWidget {
+  RedAppBar(this.title);
+  final String title;
+
+  @override
+  Size get preferredSize => const Size.fromHeight(56.0);
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: Colors.grey[300],
+        borderRadius: BorderRadius.circular(12),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.white,
+            offset: Offset(-5, -5),
+            blurRadius: 15,
+          ),
+          BoxShadow(
+            color: Colors.black.withOpacity(0.3),
+            offset: Offset(5, 5),
+            blurRadius: 15,
+          )
+        ],
+      ),
+      child: AppBar(
+        title: Text(
+          title,
+          style: TextStyle(color: Colors.white),
+        ),
+        backgroundColor: Color.fromARGB(255, 210, 15, 1),
+        //あってるかわからない↓
+        actions: [],
+        leading: IconButton(
+          icon: Icon(
+            Icons.arrow_back,
+            color: Colors.white,
+          ),
+          onPressed: () {
+            Navigator.pop(
+                context, MaterialPageRoute(builder: (context) => MyHomePage()));
+          },
+        ),
+      ),
+    );
+  }
+}
